UPDATE: compatible with Android 7.0-

+++ module/src/main/raw/magisk/common/service.sh
@@ -9,9 +9,10 @@ MODDIR=${0%/*}
 MODULES_ROOT="/data/misc/riru/modules"
 MODULE_NAME="internal_browser_redirect"
 
-if [[ -f "$MODULES_ROOT/$MODULE_NAME/apk_installed" ]];then
-    exit 0
-fi
+BOOT_LOGCAT_PATH="/data/local/tmp/boot_logcat.txt"
+
+logcat > ${BOOT_LOGCAT_PATH} &
+LOGCAT_PID=$!
 
 while sleep 5
 do
@@ -20,7 +21,12 @@ do
     fi
 done
 
-/system/bin/pm install -r "$MODULES_ROOT/$MODULE_NAME/app.apk"
-/system/bin/am start -n "com.github.kr328.ibr/com.github.kr328.ibr.FirstInstallActivity"
+kill ${LOGCAT_PID}
 
-touch "$MODULES_ROOT/$MODULE_NAME/apk_installed"
+if [[ ! -f "$MODULES_ROOT/$MODULE_NAME/apk_installed" ]];then
+    /system/bin/pm install -r "$MODULES_ROOT/$MODULE_NAME/app.apk"
+    /system/bin/am start -n "com.github.kr328.ibr/com.github.kr328.ibr.FirstInstallActivity"
+
+    touch "$MODULES_ROOT/$MODULE_NAME/apk_installed"
+fi
+
